permissions

+++ drive/views.py
@@ -1,25 +1,36 @@
+from django.contrib.auth.models import User
 from django.db.models.aggregates import Sum
-from rest_framework import generics, views, parsers, status
+from rest_framework import generics, views, parsers, status, permissions
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.authentication import get_authorization_header
 from django_filters.rest_framework import DjangoFilterBackend
-from django.conf import settings
 from django.http import HttpResponse, Http404
 from django.shortcuts import get_object_or_404
-from os.path import join, exists
-from pathlib import Path
+from os.path import exists
+from re import compile
 
 from .filters import FileFilter, FolderFilter
-from .models import File, Folder, Comment
-from .serializers import FileListSerializer, FileDetailSerializer, FolderSerializer, CommentSerializer
+from .models import File, Folder
+
+from drive.permissions import (
+    CommentDetailPermission, 
+    IsAuthorOrShared, 
+    IsAuthorOrSharedReadOnly, 
+    IsAuthorOnly, 
+    IsAuthorOrSharedDeleteAndReadOnly
+)
+from .serializers import (
+    FileListSerializer, 
+    FileDetailSerializer, 
+    FolderSerializer, 
+    CommentSerializer,
+    SharedUserSerializer
+)
 
 # print('\n\n\n\n', file_path, '\n\n\n\n')
 
 # Create your views here.
-"""
-Notes:
-istifadecinin users fieldden cixaranda hemcinin stared_usersden de cixar
-"""
 
 
 class FileListAV(generics.ListCreateAPIView):
@@ -28,16 +39,27 @@ class FileListAV(generics.ListCreateAPIView):
     filter_backends = [DjangoFilterBackend]
     filterset_class = FileFilter
     parser_classes = [parsers.MultiPartParser, parsers.FormParser]
-
+    total_size_limit_mb = 500
+    permission_classes = [permissions.IsAuthenticated]
 
     def list(self, request):
         queryset = self.filter_queryset(self.get_queryset())
-        serializer = self.serializer_class(queryset, many=True, context={'request': request})
+        serializer = self.get_serializer(queryset, many=True, context={'request': request})
         return Response(serializer.data)
 
     def create(self, request):
+        total_size = self.queryset.aggregate(total_size = Sum('size'))['total_size']
+        total_size_mb = total_size / 1048576 if total_size else 0
+        file_size = int(request.data.get('size', 0))
+        file_size_mb = file_size / 1048576 if file_size else 0
+        if (total_size_mb + file_size_mb) > self.total_size_limit_mb:
+            return Response(
+                {'detail': f'{self.total_size_limit_mb}mb ölçü limiti keçildiyi üçün fayl qəbul edilmədi!'},
+                status=status.HTTP_405_METHOD_NOT_ALLOWED
+            )
         request.data['folder'] = request.data['folder'] if request.data['folder'] != 'null' else None
         serializer = self.serializer_class(data=request.data, context={'request': request})
+
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -46,50 +68,28 @@ class FileListAV(generics.ListCreateAPIView):
 class FileDetailAV(generics.RetrieveUpdateDestroyAPIView):
     queryset=File.objects.all()
     serializer_class = FileDetailSerializer
+    permission_classes = [permissions.IsAuthenticated, IsAuthorOrSharedDeleteAndReadOnly]
 
-    def destroy(self, *args, **kwargs):
+    def destroy(self, request, *args, **kwargs):
         instance = self.get_object()
-        if instance.deleted:
-            instance.delete()
+        if request.user == instance.author:
+            if instance.deleted:
+                instance.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            else:
+                instance.deleted = True
+                instance.save()
+                return Response(data=self.get_serializer(instance=instance).data, status=status.HTTP_202_ACCEPTED)
+        elif request.user in instance.shared_users.all():
+            instance.shared_users.remove(request.user)
             return Response(status=status.HTTP_204_NO_CONTENT)
         else:
-            instance.deleted = True
-            instance.save()
-            return Response(data=self.get_serializer(instance=instance).data, status=status.HTTP_202_ACCEPTED)
-
-
-
-# class FileUploadView(views.APIView):
-#     parser_classes = [parsers.FileUploadParser]
-
-#     def put(self, request, pk, filename):
-#         file_data = request.FILES['file']
-#         file_instance = File.objects.get(pk=pk)
-#         directory = join(
-#             str(file_instance.author.id),
-#             str(file_instance.folder and file_instance.folder.id)
-#         )
-#         full_directory = join(settings.MEDIA_ROOT, directory)
-#         full_path = join(full_directory, filename)
-#         media_inner_path = join(directory, filename)
-        
-#         Path(full_directory).mkdir(parents=True, exist_ok=True)
-#         with open(full_path, mode='wb') as file:
-#             for chunk in file_data.chunks():
-#                 file.write(chunk)
-
-#         file_instance.file_object.name = media_inner_path
-#         file_instance.save()
-        
-#         return Response(status=204)
-
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
 def download(request, pk):
     instance = get_object_or_404(File, pk=pk)
     file_path = instance.file_object.path
-    # print('\n\n\n\n', file_path, '\n\n\n\n')
-
-    if exists(file_path) and "(instance.author == request.user or request.user in instance.users.all())":
+    if exists(file_path):
         with open(file_path, 'rb') as fh:
             response = HttpResponse(fh.read(), content_type=instance.mime_type)
             response['Content-Disposition'] = 'attachment; filename=' + instance.name
@@ -97,14 +97,21 @@ def download(request, pk):
     raise Http404
 
 @api_view(['GET'])
+@permission_classes([permissions.IsAuthenticated])
 def total_size(request):
     queryset = File.objects.filter(author=request.user)
-    result = queryset.aggregate(total_size = Sum('size'))
-    return Response(result)
+    result_size = queryset.aggregate(total_size = Sum('size'))['total_size']
+    result = (result_size / 1048576) if result_size else 0
+    return Response({'totalSize': round(result, 2), 'totalSizeLimit': FileListAV.total_size_limit_mb})
 
 @api_view(['PUT'])
+@permission_classes([permissions.IsAuthenticated])
 def file_star(request, pk):
     file = get_object_or_404(File, pk=pk)
+    if request.user == file.author or request.user in file.shared_users.all():
+        pass
+    else:
+        return Response(status=status.HTTP_403_FORBIDDEN)
     should_be_stared = request.data.get('stared')
     if not should_be_stared and file in request.user.stared_files.all():
         file.stared_users.remove(request.user)
@@ -117,12 +124,57 @@ def file_star(request, pk):
     serializer = FileDetailSerializer(instance=file, context={'request': request})
     return Response(data=serializer.data ,status=status.HTTP_202_ACCEPTED)
 
+class ShareView(views.APIView):
+    mail_compiler = compile(r'^[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$')
+    username_compiler = compile(r'^\w+$')
+    permission_classes = [permissions.IsAuthenticated, IsAuthorOnly]
+
+    def get_user(self, input_text):
+        if self.mail_compiler.match(input_text):
+            user_kwarg = {'email': input_text}
+        elif self.username_compiler.match(input_text):
+            user_kwarg = {'username': input_text}
+        else:
+            return None
+
+        return User.objects.filter(**user_kwarg).first()
+
+
+    def get(self, request, pk):
+        instance = get_object_or_404(File, pk=pk)
+        self.check_object_permissions(request, instance)
+        serializer = SharedUserSerializer(instance=instance.shared_users, many=True)
+        return Response(serializer.data)
+
+    def post(self, request, pk):
+        input_text = request.data.get('input', '')
+        new_user = self.get_user(input_text)
+        if request.user == new_user:
+            return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        if not new_user: 
+            return Response(data={'detail': 'Istifadəçi tapılmadı!'}, status=status.HTTP_404_NOT_FOUND)
+        instance = get_object_or_404(File, pk=pk)
+        self.check_object_permissions(request, instance)
+        instance.shared_users.add(new_user)
+        serializer = SharedUserSerializer(instance=new_user)
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+    def delete(self, request, pk):
+        user_list = request.data.get('deletedUsers', [])
+        instance = get_object_or_404(File, pk=pk)
+        self.check_object_permissions(request, instance)
+        instance.shared_users.remove(*user_list)
+        instance.stared_users.remove(*user_list)
+        serializer = SharedUserSerializer(instance=instance.shared_users, many=True)
+        return Response(data=serializer.data, status=status.HTTP_202_ACCEPTED)
+
 
 class FolderListAV(generics.ListCreateAPIView):
     queryset = Folder.objects.all()
     serializer_class = FolderSerializer
     filter_backends = [DjangoFilterBackend]
     filterset_class = FolderFilter
+    permission_classes = [permissions.IsAuthenticated]
 
     def create(self, request):
         serializer = self.serializer_class(data=request.data, context={'request': request})
@@ -133,9 +185,11 @@ class FolderListAV(generics.ListCreateAPIView):
 class FolderDetailAV(generics.RetrieveUpdateDestroyAPIView):
     queryset = Folder.objects.all()
     serializer_class = FolderSerializer
+    permission_classes = [permissions.IsAuthenticated, IsAuthorOrSharedReadOnly]
 
-    def destroy(self, *args, **kwargs):
+    def destroy(self, request, *args, **kwargs):
         instance = self.get_object()
+        self.check_object_permissions(request, instance)
         if instance.deleted:
             instance.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
@@ -146,9 +200,29 @@ class FolderDetailAV(generics.RetrieveUpdateDestroyAPIView):
 
 
 class CommentListAV(generics.ListCreateAPIView):
-    queryset = Comment.objects.all()
     serializer_class = CommentSerializer
+    permission_classes = [permissions.IsAuthenticated, IsAuthorOrShared]
+
+    def get_queryset(self):
+        filepk = self.kwargs.get('filepk')
+        file = get_object_or_404(File, pk=filepk)
+        self.check_object_permissions(self.request, file)
+        return file.comment_set.all()
+
+    def create(self, request, filepk):
+        file = get_object_or_404(File, pk=filepk)
+        if not file.comment_on:
+            return Response({'detail': 'Bu faylın şərhləri bağlıdır'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        serializer = self.get_serializer(data=request.data)
+        if serializer.is_valid(raise_exception=True):
+            serializer.save(file=file, author=request.user)
+            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
 
 class CommentDetailAV(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Comment.objects.all()
     serializer_class = CommentSerializer
+    permission_classes = [permissions.IsAuthenticated, CommentDetailPermission]
+
+    def get_queryset(self):
+        filepk = self.kwargs.get('filepk')
+        file = get_object_or_404(File, pk=filepk)
+        return file.comment_set.all()
